microservices/nifi-service/scripts/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
pandas==2.0.3
numpy==1.24.3
